make dpi double

+++ vips.h
@@ -60,7 +60,7 @@ typedef struct {
 
 typedef struct {
 	int PageNum;
-	int Dpi;
+	double Dpi;
 } Options;
 
 static unsigned long
@@ -504,7 +504,7 @@ vips_init_image (void *buf, size_t len, int imageType, VipsImage **out, Options 
 	} else if (imageType == GIF) {
 		code = vips_gifload_buffer(buf, len, out, "access", VIPS_ACCESS_RANDOM, "page", opts->PageNum, NULL);
 	} else if (imageType == PDF) {
-		code = vips_pdfload_buffer(buf, len, out, "access", VIPS_ACCESS_RANDOM, "dpi", opts->Dpi, "page", opts->PageNum, NULL);
+		code = vips_pdfload_buffer(buf, len, out, "access", VIPS_ACCESS_RANDOM, "dpi", (gdouble)opts->Dpi, "page", opts->PageNum, NULL);
 	} else if (imageType == SVG) {
 		code = vips_svgload_buffer(buf, len, out, "access", VIPS_ACCESS_RANDOM, NULL);
 #endif
